UPDATES
The main thing I added is limiting the mysqli queries in the config with
the function name of sql_query_limit. Plus the code HAS been taken from
phpbb and it is good because its states in there gpu "When we speak of
free software, we are referring to freedom, not price. Our General
Public Licenses are designed to make sure that you have the freedom to
distribute copies of free software (and charge for this service if you
wish), that you receive source code or can get it if you want it, that
you can change the software or use pieces of it in new free programs;
and that you know you can do these things." I kinda modified it use
pieces of there software.