Forums
Finally update the forums and get some progress done. This is
currentally a bugged master because there is one bug that I really want
to squash and it won't work and I need to sleep for an event on
Saturday. All i added is the cid fid and getting it into the
breadcrumbs. The bug is when you have a link like
http://localhost/pages/forums/?cid=1 it will display warnings and I want
some sleep.