Middle of Rework
I've been kind of busy this day and yesterday switching to a different
style of mysql database connection and useage. There may be bugs in this
commit and I am too lazy to test them out now. Again this is a backup to
the file so I can use them at school and any where.