Middle of Rework(p2)
This has more reworking and now I am working on a TODO list. With this
list I'll keep organise on the website progress and what I need to do
first to last. Hopefully tomarrow I get up early enough to code on this
some more. Sorry about my spelling I'm not the greatest and this is
typed really quick. Plus I'm too lazy to redo the spelling :P